Refactor organization dropdown
