+++ frontend/src/modules/organization/components/list/organization-list-table.vue
@@ -632,10 +632,18 @@
                       }"
                       class="flex justify-center"
                     >
-                      <app-organization-dropdown
-                        :organization="scope.row"
-                        @merge="isMergeDialogOpen = scope.row"
-                      />
+                      <button
+                        :id="`buttonRef-${scope.row.id}`"
+                        :ref="(el) => setActionBtnsRef(el, scope.row.id)"
+                        class="el-dropdown-link btn p-1.5 rounder-md hover:bg-gray-200 text-gray-600"
+                        type="button"
+                        @click.prevent.stop="() => onActionBtnClick(scope.row)"
+                      >
+                        <i
+                          :id="`buttonRefIcon-${scope.row.id}`"
+                          class="text-xl ri-more-fill"
+                        />
+                      </button>
                     </router-link>
                   </template>
                 </el-table-column>
@@ -663,6 +671,24 @@
         </div>
       </div>
     </div>
+    <el-popover
+      ref="OrganizationDropdownPopover"
+      placement="bottom-end"
+      popper-class="popover-dropdown"
+      :virtual-ref="actionBtnRefs[selectedActionOrganization?.id]"
+      trigger="click"
+      :visible="showOrganizationDropdownPopover"
+      virtual-triggering
+      @hide="onHide"
+    >
+      <div v-click-outside="onClickOutside">
+        <app-organization-dropdown-content
+          v-if="selectedActionOrganization"
+          :organization="selectedActionOrganization"
+          @merge="isMergeDialogOpen = selectedActionOrganization"
+        />
+      </div>
+    </el-popover>
     <app-organization-merge-dialog v-model="isMergeDialogOpen" />
   </div>
 </template>
@@ -685,10 +711,11 @@ import employeeChurnRate from '@/modules/organization/config/enrichment/employee
 import employeeGrowthRate from '@/modules/organization/config/enrichment/employeeGrowthRate';
 import revenueRange from '@/modules/organization/config/enrichment/revenueRange';
 import AppTagList from '@/modules/tag/components/tag-list.vue';
+import { ClickOutside as vClickOutside } from 'element-plus';
 import AppOrganizationIdentities from '../organization-identities.vue';
 import AppOrganizationListToolbar from './organization-list-toolbar.vue';
 import AppOrganizationName from '../organization-name.vue';
-import AppOrganizationDropdown from '../organization-dropdown.vue';
+import AppOrganizationDropdownContent from '../organization-dropdown-content.vue';
 
 const router = useRouter();
 
@@ -725,6 +752,11 @@ const tableHeaderRef = ref();
 const isScrollbarVisible = ref(false);
 const isTableHovered = ref(false);
 const isCursorDown = ref(false);
+
+const showOrganizationDropdownPopover = ref(false);
+const OrganizationDropdownPopover = ref(null);
+const actionBtnRefs = ref({});
+const selectedActionOrganization = ref(null);
 
 const pagination = computed({
   get() {
@@ -771,6 +803,34 @@ document.onmouseup = () => {
   // according to wether the mouse is hovering the table or not
   isScrollbarVisible.value = isTableHovered.value;
   isCursorDown.value = false;
+};
+
+const setActionBtnsRef = (el, id) => {
+  if (el) {
+    actionBtnRefs.value[id] = el;
+  }
+};
+
+const onActionBtnClick = (id) => {
+  showOrganizationDropdownPopover.value = !showOrganizationDropdownPopover.value;
+
+  if (selectedActionOrganization.value) {
+    setTimeout(() => {
+      selectedActionOrganization.value = null;
+    }, 200);
+  } else {
+    selectedActionOrganization.value = id;
+  }
+};
+
+const onClickOutside = (el) => {
+  if (!el.target?.id.includes('buttonRef')) {
+    showOrganizationDropdownPopover.value = false;
+
+    setTimeout(() => {
+      selectedActionOrganization.value = null;
+    }, 200);
+  }
 };
 
 function doChangeSort(sorter) {
@@ -938,5 +998,10 @@ export default {
 }
 .el-table__body {
   height: 1px;
+}
+
+.popover-dropdown {
+  padding: 0.5rem !important;
+  width: fit-content !important;
 }
 </style>
